Update to webDriver settings
To allow for flexiblity, the introduction of the ability to use any
chromium based browser to use chromedriver. Use environment variable to
access the path of the browser (SELENIUM_BROWSER_PATH) that is preferred
by the user.

Minor changes:

- Fixed structure of main function to not allow for a an error to occur
if a webdriver was not passed to the webDriver variable when trying to
close a none value.