Implement OOP
Current application structure contains a lot of redundant code, that can be refactored
and made cleaner and simpler through the use of OOP implementation. Creation of BaseQueryDriver
that emulates the base workings needed to search a given retail website for an item.

Creation of child class AmazonQueryDriver that is specific to amazon item query. This is
necessary due to Amazon searching needs for additional funcitonality that strips
unnessary information from given item's data.